backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
httpx
postgrest
realtime
storage3
supafunc
